pyproject.toml
[tool.poetry]
name = "Transformation"
version = "0.1.0"
description = ""
authors = ["Feugmo <giresse.feugmo@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
ase = "^3.22.1"
PyYAML = "^6.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
